move svelte-ignore comment extraction to parser phase
