+++ src/compiler/compile/nodes/Comment.ts
@@ -1,6 +1,5 @@
 import { TemplateNode } from '../../interfaces';
 import Component from '../Component';
-import { extract_svelte_ignore } from '../utils/extract_svelte_ignore';
 import Node from './shared/Node';
 import TemplateScope from './shared/TemplateScope';
 
@@ -12,6 +11,6 @@ export default class Comment extends Node {
 	constructor(component: Component, parent: Node, scope: TemplateScope, info: TemplateNode) {
 		super(component, parent, scope, info);
 		this.data = info.data;
-		this.ignores = extract_svelte_ignore(this.data);
+		this.ignores = info.ignores;
 	}
 }
