+++ src/compiler/utils/flatten.ts
