+++ test/parser/samples/comment-with-ignores/input.svelte
@@ -1,0 +1,1 @@
+<!-- svelte-ignore foo bar -->
